Move main-header to core/components
