+++ src/app/core/components/main-header/main-header.component.scss
