+++ src/app/core/components/main-header/main-header.module.ts
